requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
passlib[bcrypt]
python-jose
python-multipart
requests
httpx
pytest
bcrypt
websockets
